requirements.txt
# Core Dependencies (works on all platforms)
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Optional: For full face recognition (Ubuntu/Linux recommended)
# face-recognition>=1.3.0
# dlib>=19.24.0

# Database (optional, uses SQLite by default)
# psycopg2-binary>=2.9.7
# pymongo>=4.5.0
# sqlalchemy>=2.0.21

# Web Framework
fastapi==0.103.1
uvicorn[standard]==0.23.2

# Async Processing
asyncio-mqtt==0.11.1
aiofiles==23.2.1

# Notification Services
pyfcm==1.5.4
requests==2.31.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging and Monitoring
structlog==23.1.0
prometheus-client==0.17.1

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-mock==3.11.1

# Development Tools
black==23.9.1
flake8==6.1.0
mypy==1.5.1

# Machine Learning
scikit-learn==1.3.0
torch==2.0.1
torchvision==0.15.2

# Utilities
click==8.1.7
tqdm==4.66.1
python-dateutil==2.8.2
